+++ database/migrations/2020_02_19_183427_create_featured_torrents_table.php
@@ -25,9 +25,9 @@ class CreateFeaturedTorrentsTable extends Migration
     public function up()
     {
         Schema::create('featured_torrents', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id');
-            $table->integer('torrent_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('torrent_id');
             $table->nullableTimestamps();
         });
