+++ database/migrations/2020_02_19_183431_create_freeleech_tokens_table.php
@@ -25,9 +25,9 @@ class CreateFreeleechTokensTable extends Migration
     public function up()
     {
         Schema::create('freeleech_tokens', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id');
-            $table->integer('torrent_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('torrent_id');
             $table->nullableTimestamps();
         });
