+++ database/migrations/2020_02_19_183436_create_invites_table.php
@@ -25,12 +25,12 @@ class CreateInvitesTable extends Migration
     public function up()
     {
         Schema::create('invites', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
             $table->string('email');
             $table->string('code');
             $table->dateTime('expires_on')->nullable();
-            $table->integer('accepted_by')->nullable();
+            $table->unsignedBigInteger('accepted_by')->nullable();
             $table->dateTime('accepted_at')->nullable();
             $table->text('custom')->nullable();
             $table->nullableTimestamps();
