+++ database/migrations/2020_02_19_183438_create_likes_table.php
@@ -25,9 +25,9 @@ class CreateLikesTable extends Migration
     public function up()
     {
         Schema::create('likes', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id');
-            $table->integer('post_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('post_id');
             $table->boolean('like')->nullable();
             $table->boolean('dislike')->nullable();
             $table->nullableTimestamps();
