+++ database/migrations/2020_02_19_183505_create_user_notes_table.php
@@ -25,9 +25,9 @@ class CreateUserNotesTable extends Migration
     public function up()
     {
         Schema::create('user_notes', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id');
-            $table->integer('staff_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('staff_id');
             $table->text('message');
             $table->nullableTimestamps();
         });
