+++ database/migrations/2020_02_19_183509_create_wishes_table.php
@@ -25,8 +25,8 @@ class CreateWishesTable extends Migration
     public function up()
     {
         Schema::create('wishes', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
             $table->string('title');
             $table->string('imdb');
             $table->string('type');
