+++ database/migrations/2020_02_19_183513_create_clients_table.php
@@ -25,8 +25,8 @@ class CreateClientsTable extends Migration
     public function up()
     {
         Schema::create('clients', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
             $table->string('name');
             $table->string('ip');
             $table->nullableTimestamps();
