+++ database/migrations/2020_02_19_183515_create_history_table.php
@@ -25,8 +25,8 @@ class CreateHistoryTable extends Migration
     public function up()
     {
         Schema::create('history', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id');
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
             $table->string('agent')->nullable();
             $table->string('info_hash');
             $table->unsignedBigInteger('uploaded')->nullable();
