+++ database/migrations/2020_02_19_183519_create_user_audibles_table.php
@@ -25,11 +25,11 @@ class CreateUserAudiblesTable extends Migration
     public function up()
     {
         Schema::create('user_audibles', function (Blueprint $table) {
-            $table->increments('id');
-            $table->integer('user_id');
-            $table->integer('room_id')->nullable();
-            $table->integer('target_id')->nullable();
-            $table->integer('bot_id')->nullable();
+            $table->bigIncrements('id');
+            $table->unsignedBigInteger('user_id');
+            $table->unsignedBigInteger('room_id')->nullable();
+            $table->unsignedBigInteger('target_id')->nullable();
+            $table->unsignedBigInteger('bot_id')->nullable();
             $table->boolean('status')->default(0);
             $table->nullableTimestamps();
         });
